back_end/pyproject.toml
[tool.poetry]
name = "work_logger"
version = "0.1.0"
description = ""
authors = ["Ryan Doan"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.2"
django-registration = "^3.1.2"
djangorestframework = "^3.12.4"
django-rest-swagger = "^2.2.0"
PyYAML = "^5.4.1"


[tool.poetry.dev-dependencies]
black = "^21.4b1"
flake8 = "^3.9.1"
django-extensions = "^3.1.3"
ipython = "^7.23.1"
python-dotenv = "^0.17.1"
coverage = "^5.5"
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
